quickview: Mostly complete and add full unit testing.
Adding more than 100 asserts and a new unit test suit. It took
longer than expected to get to a point where is looks stable.

There is a race condition left where inserting elements seem
to corrupt the cache. It is strange because the code is single
threaded and should not be prone to such issue. It may be a small
detail, but right now I fail to see it and want to move on.

So right now the code may or may not assert at runtime depending
of luck...