scrollbar: Port the code to C++
The scrollbar code was all over the place and the 2 place that used it
had different set of bugs and fixing one would cause another one in the
other API user. While for the *current* version it could stayed in QML,
to fix other categories and lazy loading scenario, only the C++ side has
the relevant data.