abstractview: Brute force the remaining code division.
(between the model change tracking and views)

This is a very, very bad middle step in the latest row of separation of
concerns changes. It's both much slower and much uglier than before.

This commit compile and load lists and not much else. Many commits will
be needed to address the TODO/FIXME and integrate the range concept.

One nice thing is that there is now an enforced barrier between the
"view item" and "model item" concepts that cannot be bypassed due to
lazyness. It will pay off, but right now it's a mess.