abtractview: Add gears to apply the model change after a delay.
This will prevent the race condition where QML properties are added
in a random order and the delegate is set after the model. When that
happens, there is a lot of CPU cycle wasted trying to figure out what
to load.

This commit doesn't use any delay itself. There is some low hanging bug
that are easier to fix now rather than after the delay is enabled.