qmlcontext: Refactor the QML Context property handling into a class.
This commit extract more logic out of each class and group it into a
better "home". The ContextManager class now allows all the role based
optimizations to be shared across all views.

This commit does not contains the optimizations, it only contains the
refactoring part.

The commit also replace the HierarchyView class implementation with
a copy/pasted version of the QuickTreeView implementation. The old
HierarchyView lacked many shared code path and is based on an earlier
design. It was more effort to keep the old code path than replacing
them with the TreeView until more work is merged to allow the
"hierarchy" concept itself to be implemented again. For now only the
cartesian navigation accessors are implemented, but this view also
require the parent/children/siblings ones to be reliable.