contextmanager: Implement a fast metaobject generator.
This commit makes exposing the model roles to the view very, very
fast. It also adds an extension system so the views can add more
context properties. The API was designed so the model role management
is implemented as an extension itself. This reduce the code
duplication and improve readability.

To improve performance, a vTable is created in memory with the same
design as the one generated by automoc. Mapping the interal QObject
propoerty UID with the model role ID is a O(1) pointer aithmetic
operation. The only downside is that the view implementation need
to extend a class to be able to use it. The API is also very low
level and use IDs instead of names out of necessity.

This commit make the QtCore::private API required. Another commit
will make it optional again. The new system should be equally fast
if not faster than the one implemented within QtQuick, but is
"better" because it allows new views to add properties without a
massive ammount of boiler plate code.