visiblerange: Move the size hints into its own class (1/2).
This is the groundwork to move the size handling into the new
VisibleRange class. For now the code isn't really called and all
items are still loaded.

The general idea is to avoid exposing TreeTraversalItems outside
of its class, but it causes a lot of trouble. I hope solutions can
be found for each of them. If the size handling could be contained,
it would make the (final) implementation cleaner.