modeladapter: Add a new `ModelAdapter` class.
`::FlickableView` will be dropped. The main reason for the change
is because all `::FlickableView` did has already been moved elsewhere
and there is not enough left to keep its casting overhead.

On the plus side, it will allow multiple models to be displayed in the
same view. That's potentially useful.