visibleRange: Invert ownership (1/2).
The TreeTraversalReflector no longer have ranges, but rather
a range has a single TreeTraversalReflector.

It reduces flexibility, but after painful consideration is the
better way forward. The TreeTraversalReflector complexity is not
yet under control and having to manage multiple range correctly
would add even more. By contrast, having multiple
TreeTraversalReflector for the same view can mostly be used to the
same effect at the cost of more memory and CPU. However given for
the vast majority of use case there is only one, the difference is
mostly theorical.

This will allow to TreeTraversalReflector to become a  "puppet" of
the range instead of having to manage them. The tree wont have to
known about geometry anymore. The current model was a dead end if
maintainability is a goal.

This commit also adds signal tracking to the visible range instead
of being fed by the ViewBase state machine. This adds better
encapsulation and reduce the size and fragility of the public API.