visibleRange: Invert ownership (2/2).
Also replace the pure virtual `ViewBase::createItem()` by a factory
object. It's not pretty either, but prevent the need for even more
`friend` keywords to be added.

The commit breaks the `isEmpty` and `count` properties, but they were
already unreliable. It is not worth fixing right now as a better fix
will soon be available.