viewport: Ground work to enable lazy loading and recycling.
While it doesn't change much, at least now it tracks the size
of the loaded elements, the edges where more elements can be added
and where there is invisible elements.

The commit also almost complete the feedback loops that drives the
TreeTraversalItems state machine. The only reason it's not complete
is to commit before breaking /everything/ by enabling the loop to
drive the view.

So that's it, it's bug fixing time, the first iteration of the
project is complete.