GeometryCache: Complete the implementation.
The state machine was in an unfinished state since a  while. This
commit implements the methods of the state transition.

The name and scope of the actions have also been refined. It allows
the geometry cache to take into account the border "decoration"
around the index delegate. This was necessary to allow the ListView
and TreeView to display their companion widget (categories for the
former and [+] for the later) with deterministic geometry.