indexMetadata: Move into its own file.
This is the first part of a 2 commit set to encapsulate the
GeometryState and never let it be modified outside of the
IndexMetadata class.

This will allow to add an entra "layer" of logic to the GeometryCache
state machine. For example the IndexMetadata has way to know the size
and position depending on the size hint strategy.

This commit also removes dead and unnecessary code and clean up the
separation of concern. The ModelAdapter now has as a `hasSizeHints()`
property.