stateTracker: Begin to make the proximity code to a new state tracker.
For now it is incomplete, but compiles and does things. The idea is to
split what used to be TreeTraversalBase and TreeTraversalItem into two
independant classes aggregated in IndexMetadata instead of using
inheritance.

Right now the lazy loading causes "holes" where valid QModelIndex
are accidently ignored and skipped in the view. Fixing this with
the current code would add complexity. Instead of that, tracking the
state atomically should spread the complexity horizontally instead of
making the current classes an unreadable mesh of spagatti.

As for why splitting the state tracker: because I can, so I will.