viewport: Break into a dozen smaller classes.
To fix some scalability issues, this turns the SizeHintStrategy enum
into a capability based super classes. It also adds a class to
decide the optimal strategy.

Finally, it adds the skeleton of a ViewportAdapter to handle
computing the total size. This is a very early work in process,
but it's predecessor was returning broken results anyway. A lot of
work will be needed to make it work as adapters based only on the
public API, but in a sense it wasn't possible to fix all the bugs
with the current Viewport design due to hardcoded assumptions.