StateTracker::Model: Fix a stack overflow when used without flickables.
If the size of the content reflects the size of the viewport, then
inserting elements would cause a viewport resize, which would then
populate/trim causing more resizes until Q_ASSERT are hit in debug
mode or a stack overflow in release mode.

The mitigation is a hack, but until the ViewportAdapter is implemented
it is "good enough", as the code will be replaced by a safer interface.