ContextAdapter: Set a parent object with a londer lifecycle
This remove most "Rebuilding the cache because QtQuick bugs trashed
it", even if it's not the cleanest way to do it.

Those re-created contexts have some race conditions, incluring
being reparented to an object currently being deleted (corrupting
its children list).

Given most ContextAdapter classes are not QObject, there is no
good candidate for the parent and it *needs* one to prevent
QtQuick from claiming ownership.