viewItem: Actually delete.
The is rather major. The d_ptr was being deleted instead of the object.
Given the abstractitemadapter itself was no longer referenced, it was
not crashing, but it also means the QQuickItem was never "really"
deleted either. This, in turn, prevented some destructor from cleaning
the state and left invalid pointers.

It really is a ****ing miracle it took this long to cause a crash.