viewportAdapter: Add a basic implementation skeleton.
I am not quite certain how this will turn out yet, but givem at least
the sizeHint and positionHint proxies are set in stone and required
now, this can be added.

One of the big issue is that all the state machines have better
information to take decisions but cannot be exported to the public API.
A careful balance have to be found about what to expose or keep
internal.